Рефакторинг теста удаления клиента
Функция поиска клиента, подлежащего удалению, вынесена в хелпер helpers\find_del_client.py
