+++ tests/test_delete_client.py
@@ -1,22 +1,7 @@
 import allure
 from conftest import browser
 from ui.pages.manager_form import ManagerFormMethods
-
-
-def find_del_client(client_names_list):
-    name_len_sum = 0
-    del_client = ""
-    for name in client_names_list:
-        name_len_sum += len(name)
-    av_name_len = name_len_sum / len(client_names_list)
-    closest = 999
-    for name in client_names_list:
-        avdif = abs(len(name) - av_name_len)
-        if avdif < closest:
-            del_client = name
-            closest = avdif
-
-    return del_client
+from helpers.find_del_client import find_del_client
 
 
 @allure.epic("UI test")
